back/requirements.txt
Django==4.2.3
djangorestframework==3.15.1
markdown==3.6
django-filter==24.2
dj-database-url==2.0.0
gunicorn==21.2.0
wheel==0.38.1
whitenoise==6.5.0
django-cors-headers==4.3.1
requests==2.31.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
